Add files via upload
This is one of the simplest project to start with to learn about file systems in C++. The project involves a user registration process by asking USERNAME and PASSWORD. upon successful registration, a user file is created with the credentials. If the user does not exist, upon login, an error will be shown.